refactor set attributes on span
